fix(server): extract BufferImpl to reduce __before_compile__ complexity
Credo flagged cyclomatic complexity of 10 (max 9) on __before_compile__.
Extract the Buffer module body into BufferImpl macro.
